Make the CI green
* Upgrade pip before running the CI
Some TravisCI images seem to have an outdated version of pip installed,
which does not support URL-based requirements. Here we force-upgrade
pip to the latest version in the `before_install` stanza as a
workaround.

* Ignore missing docstrings for now
We should have these in the future, but viskit is an application and
not a library, so documentation is less important. I'm not going to
force the CI to red for missing docstrings (yet)

* Remove custom flake8 stage from .pre-commit-config.yml
This was copied from rlworkgroup/garage, but is not necessary in this
codebase

* Clean up flake8 violations and double-quote strings in core.py

* Add coverage configuration to setup.cfg

* Add a unit tests so that coverage works