Create a special deploy stage in TravisCI
This PR creates a special deploy stage in TravisCI which is only
executed on tag builds *after* all testing stages have completed. This
is necessary because we have more than one test job (Python versions),
but we only want to deploy to PyPI after a successful tag build.
Deploying the same file to PyPI more than once is an error.